Messaging Functional, Refactor
Messages are being saved to the db on send, still some kind of websocket connection needed for realtime stuff.

Small refactor as seperate livewire files was getting to be a pain as dispatch of livewire funcs/methods needs to happen across the sidelist and chat areas, so having them in one parent livewire component makes more sense.