Workaround for Messages not showing up in other chats
while using livewire and echo, it is well documented online that it is not possible to re-set the active listening channel. This is an incredibly frustrating thing to only have discovered now, so a workaround has been made which just determines whether or not the message is for this chat based on the assigned conversation_id in the broadcast.

Not ideal, I'm not happy with this but it will do for now.