Improve System Messages
System messages are now clear about who performed an action, what the action is, and who it affects. It also has clickable names to view the profiles of the actioning user and the affected user, where applicable.

Some limited support for deletion of users here as this will be implemented in the future