Message Dropdown for Deletion
This was actually so annoying, I was using flowbite but its JS wouldn't keep up to date as livewire changed the DOM, so had to build it with alpine. Seems to work well though, I like the loading indicator. This will be used for sending messages in the future as well

Supports the addition of Editing messages, and other features for individual messages too like reporting or anything else I want to add in the future.