Variety of fixes and improvements
Icon for Edit button on popup menu, matches delete button nicely.

Moved edit message input to a component for reuse.

Up arrow to edit last user message requires the user to be focused on the main chat area input, as some users may use the up arrows to scroll up the chat and this reduces accessibility if not done.