Remove Delete User Event
This is causing some kind of race condition between the session being terminated and the websocket event listener firing causing a 419 error. The only thing this will break is when a user deletes their account, their user profile will 404 until the view gets a livewire refresh for any users looking at the chat.